Extract field types into separate module
